requirements.txt
ambient==0.2.1
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
packaging==24.2
pip-review==1.3.0
pyserial==3.5
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.3.0
websockets==15.0
wsproto==1.2.0
